# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: SheCanCODE Attendance & Student Management System
  description: >
    API for authentication, administration (cohorts & programs) and the student
    lifecycle.

    Auth flow:
      1. An ADMIN (seeded) creates TRAINER accounts via POST /api/v1/auth/register.
      2. An ADMIN enrols a STUDENT with just an email + program via
         POST /api/v1/students/{cohortNumber}. This creates a disabled account and
         emails the student a one-time activation link.
      3. The student activates and sets a password via POST /api/v1/auth/activate,
         which returns a JWT.
      4. The student completes their profile via PUT /api/v1/students/me/profile.
  version: 2.0.0
servers:
  - url: http://localhost:8081
...
paths:
...
  /api/v1/trainers:
    post:
      tags: [Trainers]
      summary: Invite a trainer (ADMIN only)
      description: >
        Creates a disabled TRAINER account and emails an activation link. The trainer
        sets their own password on activation.
      operationId: inviteTrainer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainerInviteRequest'
      responses:
        '201':
          description: Trainer invited; activation email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitedAccountResponse'
        '400':
          description: Validation failed
        '401':
          description: Missing or invalid JWT
        '403':
          description: Caller is not an ADMIN
        '409':
          description: A user with this email already exists
        '502':
          description: Invitation email could not be sent

  # ─── COHORT MANAGEMENT ─────────────────────────────────────────────────────
  /api/v1/cohort:
    post:
      tags: [Administration]
      summary: Create a new cohort (ADMIN only)
      operationId: createCohort
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CohortRequestDao'
      responses:
        '201':
          description: Cohort created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CohortResponseDao'
        '401':
          description: Missing or invalid JWT
        '403':
          description: Caller is not an ADMIN
    get:
      tags: [Administration]
      summary: List all cohorts (ADMIN only)
      operationId: getAllCohorts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CohortResponseDao'

  # ─── PROGRAM MANAGEMENT ────────────────────────────────────────────────────
  /api/v1/cohorts/{cohortNumber}/program:
    post:
      tags: [Administration]
      summary: Create a new program within a cohort (ADMIN only)
      operationId: createProgram
      security:
        - bearerAuth: []
      parameters:
        - name: cohortNumber
          in: path
          required: true
          schema:
            type: string
            example: "Cohort-18"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgramRequestDao'
      responses:
        '201':
          description: Program created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramResponseDao'
        '401':
          description: Missing or invalid JWT
        '403':
          description: Caller is not an ADMIN

  # ─── STUDENT ENROLMENT & PROFILE ───────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TrainerInviteRequest:
      type: object
      required: [email, fullName]
      properties:
        email:
          type: string
          format: email
          example: "trainer.jane@example.com"
        fullName:
          type: string
          example: "Jane Umutoni"
    InvitedAccountResponse:
      type: object
      properties:
        userId: { type: string, format: uuid }
        email: { type: string }
        fullName: { type: string }
        role: { type: string, enum: [ADMIN, TRAINER, STUDENT] }
        accountStatus: { type: string, enum: [INVITED, PROFILE_INCOMPLETE, PROFILE_COMPLETE, ACTIVE] }
...
    CohortRequestDao:
      type: object
      required: [cohortNumber, startDate, endDate]
      properties:
        cohortNumber:
          type: string
          example: "Cohort-18"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    CohortResponseDao:
      type: object
      properties:
        id: { type: string, format: uuid }
        cohortNumber: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
    ProgramRequestDao:
      type: object
      required: [programName, programDuration, programStartDate, programEndDate]
      properties:
        programName:
          type: string
          example: "BACKEND"
        programDuration:
          type: integer
          example: 180
        programStartDate:
          type: string
          format: date
        programEndDate:
          type: string
          format: date
    ProgramResponseDao:
      type: object
      properties:
        id: { type: string, format: uuid }
        programName: { type: string }
        programDuration: { type: integer }
        programStartDate: { type: string, format: date }
        programEndDate: { type: string, format: date }
        cohortNumber: { type: string }

    # ─── Students ─────────────────────────────────────────────────────────────
...